Move repeated chart factories to pages
