+++ vizro-core/examples/_chart-gallery/pages/deviation.py
@@ -1,3 +1,5 @@
-butterfly = "butterfly"
+from pages._factories import butterfly_factory
+
+butterfly = butterfly_factory("deviation")
 
 pages = [butterfly]
